Update `Collection` nomenclature to be more explicit
This updates the nomenclature of `Collection` to use the more explicit nomenclature of `Contract` as there are often contracts that have multiple different collections/projects within them (e.g. Art Blocks, various partner platforms powered by the Art Blocks Engine, Superrare, Foundation, Niftygateway), and while I don't think it makes sense for this registry to be aware of the nuances of how various contracts handle the abstraction of "collection" or "project", I do think that the nomenclature used by the registry should be as explicit here as possible to avoid any opportunity for developer / user error due to confusion.