Subdelegations, airdrop example is now a subdelegation example, up to 30% less gas for delegate methods (#36)
DelegateRegistry.sol changes
Replaced OpenZeppelin enumerable sets with bytes32[] array in vault and delegate hash mappings

Hashes for a particular vault / delegate are now pushed to a bytes32[] array in their hash mappings instead of adding (and removing) them from a set. This reduces gas by > 40k every time a delegation is added to the registry due to the storage writes in the enumerable set backend. Suppose that a vault / delegate accumulates 500 delegation hashes in their array, the increased computation of the getDelegationsFor methods is well within acceptable gas limits for most RPC providers.

Consumable methods are unaffected as they now use the _delegations mapping only to establish whether a delegation is valid and enabled / disabled.
Changed bytes to bytes32[6] in _delegations mapping

Reduces > 20k gas every time a delegation is added since variable sized storage data also need to store their length.
Updated delegate methods to write minimal storage and removed _setDelegationValues

Delegate methods now only write data if it exists in their type and does not write rights to storage if it's the default "". _computeDelegationLocation(hash) is used to obtain the storage location for a given hash, and writeDelegation(location, StoragePositions.delegate, data); is used to write particular delegation data. For example, this saves > 80k gas for a delegateForAll call with all rights. Storage positions of data types are standardised with the new StoragePositions struct. _setDelegationValues has been removed since these changes make it redundant.
Delegation type now encoded in the last byte of the hash

This solves both the problem of collisions and being able to identify the type of delegation without writing its type to storage.
Check methods now support subdelegations

All check methods now accept a bytes32 rights parameter (previously data). If this parameter is not the default "", the check will bubble through the default case and then the rights case unless it returns true before exhausting the search (in the case of erc721, contract, all) or unless it returns type(uint256).max before exhausting the search (in the erc20, erc1155 cases). The _loadDelegationAddress(location, StoragePositions.vault) and _loadDelegationUint(location, StoragePositions.balance) methods are used to load vault and balance values directly from storage. A check on vault is now used to determine if the hash exists and is valid (since it is set to zero when disabled, and set to vault when enabled).
Removed occurrences of IDelegationRegistry.[object]

These objects are inherited explicitly anyway, and omitting IDelegationRegistry from these object references increases readability of the code.
Created Consumable and Private methods sections

Mostly a code refactor to keep the write and read sections clean. Added consumable as a section since we have distinct classes of on chain functions that can be consumed on and off chain.
IDelegateRegistry.sol changes

    Removed NONE from DelegationType enum as it's never used.
    Added StoragePositions enum to standardise storage position of delegation data.
    Renamed DelegationInfo to Delegation struct to be used for returning an arbitrary delegation.
    Created BatchDelegation struct to be used for batch delegations

Airdrop Example Updated to a Subdelegation Example

The airdrop example now accepts sub delegations with rights = "airdrop" in addition to default delegations, configured with a acceptableRight variable in the airdrop contract.
New tests

    Fuzz test for supportsInterface method
    testVultEnumerationGas tests an edge case where a vault has enabled 500 delegations.